move function to reportUtils
